requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
streamlit>=1.39.0
python-dotenv>=1.0.0
